functions/requirements.txt
fastapi
uvicorn
firebase-functions
firebase-admin
a2wsgi
python-dotenv
google-cloud-secret-manager 